requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
batchgenerators==0.25
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dill==0.3.6
EditorConfig==0.12.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
google-auth==2.17.1
google-auth-oauthlib==1.0.0
googledrivedownloader==0.4
greenlet==2.0.2
grpcio==1.53.0
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.9
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
linecache2==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.2
MedPy==0.4.0
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numpy==1.24.3
numpy-stl==3.0.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.2
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.4.3
pathos==0.3.0
Pillow==9.5.0
pip==23.0.1
pox==0.3.2
ppft==1.7.6.6
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyDeprecate==0.3.2
pydicom==2.4.1
pylidc==0.2.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.7.0
pytorch-lightning==1.7.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
setuptools==65.6.3
SimpleITK==2.2.1
six==1.16.0
SQLAlchemy==2.0.19
stl-to-voxel==0.9.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.3.21
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.64.0
traceback2==1.4.0
typing_extensions==4.8.0
unittest2==1.1.0
urllib3==1.26.15
wandb==0.16.6
Werkzeug==2.2.3
wheel==0.38.4
yarl==1.8.2
